refac: pastas modularizadas
