+++ .gitignore
@@ -47,3 +47,4 @@ android/app/.cxx/
 /android/key.properties
 upload-keystore.jks
 macos/Flutter/GeneratedPluginRegistrant.swift
+android/app/google-services.json
