--- lib/app/bindings/login_bindings.dart
@@ -1,9 +1,0 @@
-import 'package:get/get.dart';
-import 'package:redescomunicacionais/app/controller/login_controller.dart';
-
-class LoginBinding implements Bindings {
-  @override
-  void dependencies() {
-    Get.lazyPut<LoginController>(() => LoginController());
-  }
-}
