--- lib/app/bindings/user_bindings.dart
@@ -1,9 +1,0 @@
-import 'package:get/get.dart';
-import 'package:redescomunicacionais/app/controller/user_controller.dart';
-
-class UserBinding implements Bindings {
-  @override
-  void dependencies() {
-    Get.lazyPut<UserController>(() => UserController());
-  }
-}
