--- lib/app/data/repository/login_repository.dart
@@ -1,30 +1,0 @@
-import 'package:redescomunicacionais/app/data/model/user_model.dart';
-import 'package:redescomunicacionais/app/data/provider/sign_in.dart';
-
-class LoginRepository {
-  final SignInService signInService = SignInService();
-
-  Future<UserModel?> signInGoogle() {
-    return signInService.signInGoogle();
-  }
-
-  Future<UserModel?> trySignInGoogle() {
-    return signInService.trySignInGoogle();
-  }
-
-  logoutGoogle() {
-    signInService.logoutGoogle();
-  }
-
-  Future<UserModel?> signInMicrosoft() async {
-    return signInService.signInMicrosoft();
-  }
-
-  Future<UserModel?> trySignInMicrosoft() {
-    return signInService.trySignInMicrosoft();
-  }
-
-  logoutMicrosoft() {
-    signInService.logoutMicrosoft();
-  }
-}
