+++ lib/app/modules/dashboard/bindings/home_bindings.dart
@@ -1,0 +1,10 @@
+import 'package:get/get.dart';
+import 'package:redescomunicacionais/app/modules/dashboard/controller/home_controller.dart';
+
+class HomeBinding implements Bindings {
+@override
+void dependencies() {
+  Get.lazyPut<HomeController>(() => HomeController());
+     
+  }
+}
