+++ lib/app/modules/login/bindings/login_bindings.dart
@@ -1,0 +1,9 @@
+import 'package:get/get.dart';
+import 'package:redescomunicacionais/app/modules/login/controller/login_controller.dart';
+
+class LoginBinding implements Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<LoginController>(() => LoginController());
+  }
+}
