+++ lib/app/modules/news/data/model/news_model.dart
