+++ lib/app/modules/user/bindings/user_bindings.dart
@@ -1,0 +1,9 @@
+import 'package:get/get.dart';
+import 'package:redescomunicacionais/app/modules/user/controller/user_controller.dart';
+
+class UserBinding implements Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<UserController>(() => UserController());
+  }
+}
