+++ lib/app/utils/components/icon_base64.dart
