+++ lib/main.dart
@@ -7,7 +7,7 @@ import 'package:redescomunicacionais/app/controller/version_controller.dart';
 import 'package:redescomunicacionais/app/routes/app_pages.dart';
 import 'package:redescomunicacionais/app/routes/app_routes.dart';
 import 'package:redescomunicacionais/app/services/hive_service.dart';
-import 'package:redescomunicacionais/app/ui/theme/app_theme.dart';
+import 'package:redescomunicacionais/app/utils/theme/app_theme.dart';
 import 'package:redescomunicacionais/firebase_options.dart';
 
 Future<void> main() async {
